fixed comments from PR#4
