+++ frontend/src/app/login/register/page.tsx
@@ -10,10 +10,10 @@ export default function Register() {
       <div className="card grid gap-4 p-6">
         <p className="text-subtext">Create an account to start donating or claiming food.</p>
         <div className="flex flex-wrap gap-2">
-          <SignUpButton mode="modal" forceRedirectUrl="login/registerdonor">
+          <SignUpButton mode="modal" forceRedirectUrl="/login/register/registerdonor">
             <button type="button" className="btn-primary">Register as Donor</button>
           </SignUpButton>
-          <SignUpButton mode="modal" forceRedirectUrl="login/registerreciver">
+          <SignUpButton mode="modal" forceRedirectUrl="/login/register/registerreciver">
             <button type="button" className="btn-primary">Register as Receiver</button>
           </SignUpButton>
         </div>
