+++ frontend/src/app/login/register/registerdonor/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useEffect, useMemo, useState } from "react";
 import { useAuth, useUser } from "@clerk/nextjs";
-import { useMutation } from "convex/react";
+import { useMutation, useQuery } from "convex/react";
 import { api } from "@/convexApi";
 import { useRouter } from "next/navigation";
 import Access from "@/components/Access/Access";
@@ -11,6 +11,7 @@ export default function RegisterDonater() {
   const { userId } = useAuth();
   const { user } = useUser();
   const registerDonor = useMutation(api.functions.createUser.registerDonor);
+  const donors = useQuery(api.functions.createUser.listDonors, {});
 
   const [firstName, setFirstName] = useState("");
   const [lastName, setLastName] = useState("");
@@ -19,6 +20,8 @@ export default function RegisterDonater() {
   const [businessName, setBusinessName] = useState("");
   const [businessEmail, setBusinessEmail] = useState("");
   const [businessPhone, setBusinessPhone] = useState("");
+  const [selectedDonorId, setSelectedDonorId] = useState<string>("");
+  const NEW_BUSINESS_VALUE = "__new_business__";
   const [submitting, setSubmitting] = useState(false);
   const [error, setError] = useState<string | null>(null);
   const [completed, setCompleted] = useState(false);
@@ -45,16 +48,11 @@ export default function RegisterDonater() {
   }, [completed]);
 
   const allValid = useMemo(() => {
-    return (
-      firstName.trim() &&
-      lastName.trim() &&
-      phone.trim() &&
-      address.trim() &&
-      businessName.trim() &&
-      businessEmail.trim() &&
-      businessPhone.trim()
-    );
-  }, [firstName, lastName, phone, address, businessName, businessEmail, businessPhone]);
+    const hasUser = firstName.trim() && lastName.trim() && phone.trim();
+    if (!hasUser) return false;
+    if (selectedDonorId && selectedDonorId !== NEW_BUSINESS_VALUE) return true;
+    return businessName.trim() && businessEmail.trim() && businessPhone.trim() && address.trim();
+  }, [firstName, lastName, phone, selectedDonorId, businessName, businessEmail, businessPhone, address]);
 
   const onSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
@@ -63,19 +61,29 @@ export default function RegisterDonater() {
       setError("You must be signed in to register.");
       return;
     }
-    if (!allValid) return;
+    if (!allValid && !selectedDonorId) return;
     try {
       setSubmitting(true);
-      await registerDonor({
-        clerk_id: userId,
-        first_name: firstName,
-        last_name: lastName,
-        phone,
-        address,
-        business_name: businessName,
-        business_email: businessEmail,
-        business_phone: businessPhone,
-      });
+      if (selectedDonorId && selectedDonorId !== NEW_BUSINESS_VALUE) {
+        await registerDonor({
+          clerk_id: userId,
+          first_name: firstName,
+          last_name: lastName,
+          phone,
+          donors_id: selectedDonorId as any,
+        } as any);
+      } else {
+        await registerDonor({
+          clerk_id: userId,
+          first_name: firstName,
+          last_name: lastName,
+          phone,
+          address,
+          business_name: businessName,
+          business_email: businessEmail,
+          business_phone: businessPhone,
+        } as any);
+      }
       setCompleted(true);
       router.replace("/dashboard");
     } catch (err) {
@@ -95,7 +103,7 @@ export default function RegisterDonater() {
   }, [user]);
 
   return (
-    <Access requireAuth requireUnregistered>
+    <Access requireAuth requireUnregistered redirectIfRegisteredTo="/dashboard">
       <section className="grid gap-4 max-w-md">
         <h2 className="text-2xl font-semibold">Register as Donor</h2>
         <div className="card grid gap-4 p-6">
@@ -116,21 +124,38 @@ export default function RegisterDonater() {
               <input className="input" value={phone} onChange={e=>setPhone(e.target.value)} />
             </label>
             <label className="grid gap-1">
-              <span className="label">Address</span>
-              <input className="input" value={address} onChange={e=>setAddress(e.target.value)} />
+              <span className="label">Organization</span>
+              <select className="input" value={selectedDonorId} onChange={e=>setSelectedDonorId(e.target.value)}>
+                <option value="">Select organization…</option>
+                <option value={NEW_BUSINESS_VALUE}>+ Add new business</option>
+                {(donors || []).map((d: { _id: string; name: string }) => (
+                  <option key={d._id} value={d._id as string}>{d.name}</option>
+                ))}
+              </select>
+              <span className="text-xs text-subtext">Choose an existing organization or add a new one.</span>
             </label>
-            <label className="grid gap-1">
-              <span className="label">Business name</span>
-              <input className="input" value={businessName} onChange={e=>setBusinessName(e.target.value)} />
-            </label>
-            <label className="grid gap-1">
-              <span className="label">Business email</span>
-              <input type="email" className="input" value={businessEmail} onChange={e=>setBusinessEmail(e.target.value)} />
-            </label>
-            <label className="grid gap-1">
-              <span className="label">Business phone</span>
-              <input className="input" value={businessPhone} onChange={e=>setBusinessPhone(e.target.value)} />
-            </label>
+            {selectedDonorId === NEW_BUSINESS_VALUE && (
+              <>
+                <label className="grid gap-1">
+                  <span className="label">Business name</span>
+                  <input className="input" value={businessName} onChange={e=>setBusinessName(e.target.value)} />
+                </label>
+                <div className="grid grid-cols-2 gap-3">
+                  <label className="grid gap-1">
+                    <span className="label">Business email</span>
+                    <input type="email" className="input" value={businessEmail} onChange={e=>setBusinessEmail(e.target.value)} />
+                  </label>
+                  <label className="grid gap-1">
+                    <span className="label">Business phone</span>
+                    <input className="input" value={businessPhone} onChange={e=>setBusinessPhone(e.target.value)} />
+                  </label>
+                </div>
+                <label className="grid gap-1">
+                  <span className="label">Business address</span>
+                  <input className="input" value={address} onChange={e=>setAddress(e.target.value)} />
+                </label>
+              </>
+            )}
             {error && <p className="text-red-500 text-sm">{error}</p>}
             <button type="submit" className="btn-primary" disabled={!allValid || submitting}>
               {submitting ? "Registering..." : "Register"}
